py/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
greenlet==1.1.1
idna==3.2
lxml==4.6.3
mysqlclient==2.0.3
python-dateutil==2.8.2
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
urllib3==1.26.6
